Remove unnecessary copy buffer to parse types

+++ S7.Net/Types/Boolean.cs
@@ -10,7 +10,7 @@
         /// </summary>
         public static bool GetValue(byte value, int bit)
         {
-            return (((int)value & (1 << bit)) != 0);
+            return (value & (1 << bit)) != 0;
         }
 
         /// <summary>
